+++ apprise/config/__init__.py
@@ -23,40 +23,70 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import sys
 import six
+import re
 
-from .ConfigHTTP import ConfigHTTP
-from .ConfigFile import ConfigFile
+from os import listdir
+from os.path import dirname
+from os.path import abspath
 
 # Maintains a mapping of all of the configuration services
 SCHEMA_MAP = {}
 
-
-__all__ = [
-    # Configuration Services
-    'ConfigFile', 'ConfigHTTP',
-]
+__all__ = []
 
 
 # Load our Lookup Matrix
-def __load_matrix():
+def __load_matrix(path=abspath(dirname(__file__)), name='apprise.config'):
     """
     Dynamically load our schema map; this allows us to gracefully
     skip over modules we simply don't have the dependencies for.
 
     """
+    # Used for the detection of additional Notify Services objects
+    # The .py extension is optional as we support loading directories too
+    module_re = re.compile(r'^(?P<name>Config[a-z0-9]+)(\.py)?$', re.I)
 
-    thismodule = sys.modules[__name__]
+    for f in listdir(path):
+        match = module_re.match(f)
+        if not match:
+            # keep going
+            continue
 
-    # to add it's mapping to our hash table
-    for entry in dir(thismodule):
+        # Store our notification/plugin name:
+        plugin_name = match.group('name')
+        try:
+            module = __import__(
+                '{}.{}'.format(name, plugin_name),
+                globals(), locals(),
+                fromlist=[plugin_name])
+
+        except ImportError:
+            # No problem, we can't use this object
+            continue
+
+        if not hasattr(module, plugin_name):
+            # Not a library we can load as it doesn't follow the simple rule
+            # that the class must bear the same name as the notification
+            # file itself.
+            continue
 
         # Get our plugin
-        plugin = getattr(thismodule, entry)
-        if not hasattr(plugin, 'app_id'):  # pragma: no branch
+        plugin = getattr(module, plugin_name)
+        if not hasattr(plugin, 'app_id'):
             # Filter out non-notification modules
             continue
+
+        elif plugin_name in __all__:
+            # we're already handling this object
+            continue
+
+        # Add our module name to our __all__
+        __all__.append(plugin_name)
+
+        # Ensure we provide the class as the reference to this directory and
+        # not the module:
+        globals()[plugin_name] = plugin
 
         # Load protocol(s) if defined
         proto = getattr(plugin, 'protocol', None)
@@ -82,6 +112,8 @@ def __load_matrix():
                 if p not in SCHEMA_MAP:
                     SCHEMA_MAP[p] = plugin
 
+    return SCHEMA_MAP
 
-# Dynamically build our module
+
+# Dynamically build our schema base
 __load_matrix()
