+++ apprise/plugins/__init__.py
@@ -23,43 +23,23 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import sys
 import six
+import re
 
+from os import listdir
+from os.path import dirname
+from os.path import abspath
+
+# Used for testing
 from . import NotifyEmail as NotifyEmailBase
 
-from .NotifyBoxcar import NotifyBoxcar
-from .NotifyDBus import NotifyDBus
-from .NotifyDiscord import NotifyDiscord
-from .NotifyEmail import NotifyEmail
-from .NotifyEmby import NotifyEmby
-from .NotifyFaast import NotifyFaast
-from .NotifyGotify import NotifyGotify
-from .NotifyGrowl.NotifyGrowl import NotifyGrowl
-from .NotifyGnome import NotifyGnome
-from .NotifyIFTTT import NotifyIFTTT
-from .NotifyJoin import NotifyJoin
-from .NotifyJSON import NotifyJSON
-from .NotifyMatrix import NotifyMatrix
-from .NotifyMatterMost import NotifyMatterMost
-from .NotifyProwl import NotifyProwl
-from .NotifyPushed import NotifyPushed
-from .NotifyPushBullet import NotifyPushBullet
-from .NotifyPushjet.NotifyPushjet import NotifyPushjet
-from .NotifyPushover import NotifyPushover
-from .NotifyRocketChat import NotifyRocketChat
-from .NotifyRyver import NotifyRyver
-from .NotifySlack import NotifySlack
-from .NotifySNS import NotifySNS
-from .NotifyTelegram import NotifyTelegram
-from .NotifyTwitter.NotifyTwitter import NotifyTwitter
-from .NotifyXBMC import NotifyXBMC
-from .NotifyXML import NotifyXML
-from .NotifyWindows import NotifyWindows
-
+# Required until re-factored into base code
 from .NotifyPushjet import pushjet
 from .NotifyGrowl import gntp
 from .NotifyTwitter import tweepy
+
+# NotifyBase object is passed in as a module not class
+from . import NotifyBase
 
 from ..common import NotifyImageSize
 from ..common import NOTIFY_IMAGE_SIZES
@@ -69,21 +49,12 @@ from ..common import NOTIFY_TYPES
 # Maintains a mapping of all of the Notification services
 SCHEMA_MAP = {}
 
-
 __all__ = [
-    # Notification Services
-    'NotifyBoxcar', 'NotifyDBus', 'NotifyEmail', 'NotifyEmby', 'NotifyDiscord',
-    'NotifyFaast', 'NotifyGnome', 'NotifyGotify', 'NotifyGrowl', 'NotifyIFTTT',
-    'NotifyJoin', 'NotifyJSON', 'NotifyMatrix', 'NotifyMatterMost',
-    'NotifyProwl', 'NotifyPushed', 'NotifyPushBullet', 'NotifyPushjet',
-    'NotifyPushover', 'NotifyRocketChat', 'NotifyRyver', 'NotifySlack',
-    'NotifySNS', 'NotifyTwitter', 'NotifyTelegram', 'NotifyXBMC',
-    'NotifyXML', 'NotifyWindows',
-
     # Reference
     'NotifyImageSize', 'NOTIFY_IMAGE_SIZES', 'NotifyType', 'NOTIFY_TYPES',
+    'NotifyBase',
 
-    # NotifyEmail Base References (used for Testing)
+    # NotifyEmail Base Module (used for NotifyEmail testing)
     'NotifyEmailBase',
 
     # gntp (used for NotifyGrowl Testing)
@@ -98,23 +69,56 @@ __all__ = [
 
 
 # Load our Lookup Matrix
-def __load_matrix():
+def __load_matrix(path=abspath(dirname(__file__)), name='apprise.plugins'):
     """
     Dynamically load our schema map; this allows us to gracefully
     skip over modules we simply don't have the dependencies for.
 
     """
+    # Used for the detection of additional Notify Services objects
+    # The .py extension is optional as we support loading directories too
+    module_re = re.compile(r'^(?P<name>Notify[a-z0-9]+)(\.py)?$', re.I)
 
-    thismodule = sys.modules[__name__]
+    for f in listdir(path):
+        match = module_re.match(f)
+        if not match:
+            # keep going
+            continue
 
-    # to add it's mapping to our hash table
-    for entry in dir(thismodule):
+        # Store our notification/plugin name:
+        plugin_name = match.group('name')
+        try:
+            module = __import__(
+                '{}.{}'.format(name, plugin_name),
+                globals(), locals(),
+                fromlist=[plugin_name])
+
+        except ImportError:
+            # No problem, we can't use this object
+            continue
+
+        if not hasattr(module, plugin_name):
+            # Not a library we can load as it doesn't follow the simple rule
+            # that the class must bear the same name as the notification
+            # file itself.
+            continue
 
         # Get our plugin
-        plugin = getattr(thismodule, entry)
-        if not hasattr(plugin, 'app_id'):  # pragma: no branch
+        plugin = getattr(module, plugin_name)
+        if not hasattr(plugin, 'app_id'):
             # Filter out non-notification modules
             continue
+
+        elif plugin_name in __all__:
+            # we're already handling this object
+            continue
+
+        # Add our module name to our __all__
+        __all__.append(plugin_name)
+
+        # Ensure we provide the class as the reference to this directory and
+        # not the module:
+        globals()[plugin_name] = plugin
 
         # Load protocol(s) if defined
         proto = getattr(plugin, 'protocol', None)
@@ -140,6 +144,8 @@ def __load_matrix():
                 if p not in SCHEMA_MAP:
                     SCHEMA_MAP[p] = plugin
 
+    return SCHEMA_MAP
 
-# Dynamically build our module
+
+# Dynamically build our schema base
 __load_matrix()
