+++ test/test_rest_plugins.py
@@ -1610,8 +1610,11 @@ def test_rest_plugins(mock_post, mock_get):
                 url, asset=asset, suppress_exceptions=False)
 
             if obj is None:
-                # We're done (assuming this is what we were expecting)
-                assert instance is None
+                if instance is not None:
+                    # We're done (assuming this is what we were expecting)
+                    print("{} didn't instantiate itself "
+                          "(we expected it to)".format(url))
+                    assert False
                 continue
 
             if instance is None:
